requirements.txt
futu_api==9.4.5408
pandas==2.1.4
streamlit==1.52.2
PyYAML==6.0.3
